+++ NEWS.md
@@ -1,4 +1,4 @@
-# pipload (development version)
+# pipload 0.1.14
 
 * make functions consistent with output from `suyrvey_id_to_vars()`